[C++] Added ground texture to PlayState
Also, updated menu font and unluckily tried to add shadows to game

+++ include/entities/Enemy.h
@@ -14,10 +14,10 @@ public:
   bool shoot();
   Ogre::Vector3 getPosition();
   Ogre::SceneNode * getSceneNode();
-  
+
 private:
   Ogre::Vector3 _position;
   Ogre::SceneNode * _node;
   Ogre::SceneManager * _sceneMgr;
-  float getRandomXPosition();
+  double getRandomXPosition();
 };
